[sdk] Move type check utils to /utils
