+++ packages/sdk/CHANGELOG.md
@@ -1,7 +1,7 @@
 ### Unreleased
 
 - YYYY-MM-DD **BREAKING?** -- description
-
+- 2021-04-23 -- Moves type check utils to `/utils`
 - 2021-04-23 -- Move `createVoucher` to `/utils` to resolve circular dependency in `interaction`
 - 2021-04-22 -- Adds `wallet-utils` `validateSignableTransaction` support for wallets to validate Signable payload
 - 2021-04-21 -- Removes **Deprecated** `params`, `buildParams`
